Feature/fix all project with transfer requests

+++ app/controllers/ctes_controller.rb
@@ -4,9 +4,7 @@ class CtesController < UsersController
   before_action :set_cte, only: %w[destroy show]
 
   def index
-    @q = Cte.includes([client: :address])
-            .includes(:truckload)
-            .includes([user: :person])
+    @q = Cte.includes([client: :address], :truckload)
             .accessible_by(current_ability)
             .page(params[:page])
             .ransack(params[:q])
@@ -15,25 +13,31 @@ class CtesController < UsersController
   end
 
   def destroy
-    if @cte.destroy
-      redirect_to ctes_path
-      flash[:success] = 'CT-e excluído com sucesso'
+    can_destroy_cte = true if current_user.roles.kind_masters.present? ||
+                              Cte.find(params[:id]).truckload.user == current_user
+
+    if can_destroy_cte
+      if @cte.destroy
+        redirect_to ctes_path
+        flash[:success] = 'CT-e excluído com sucesso'
+      else
+        render :index
+      end
     else
-      render :index
+      redirect_to(ctes_path)
+      flash[:danger] = 'Você não pode deletar o CT-e de terceiros.'
     end
   end
 
   private
 
   def set_cte
-    if current_user.cte_ids.include?(Cte.find(params[:id]).id) ||
-       current_user.roles.kind_masters.present? ||
-       (current_user.roles.kind_owners.present? &&
-        current_user.enterprise_id == Cte.find(params[:id]).enterprise_id)
+    if current_user.roles.kind_masters.present? ||
+       current_user.enterprise_id == Cte.find(params[:id]).enterprise_id
       @cte = Cte.find(params[:id])
     else
       redirect_to root_path
-      flash[:danger] = 'Você não tem permissão para manipular este CT-e.'
+      flash[:danger] = 'Você não possui permissão para manipular este CT-e.'
     end
   end
 end
